Need to fix mismatch

+++ app/src/main/java/com/example/knocker/model/ContactDetails.kt
@@ -7,8 +7,8 @@ import androidx.room.*
         childColumns = arrayOf("id_contact"),
         onDelete = ForeignKey.CASCADE)))
 data class ContactDetails(
-        @ColumnInfo(name = "id") @PrimaryKey(autoGenerate = true) val id: Long?,    //id du contact
-        @ColumnInfo(name = "id_contact") val idContact: Long?,                      //id du contact
+        @ColumnInfo(name = "id") @PrimaryKey(autoGenerate = true) val id: Int?,    //id du contact
+        @ColumnInfo(name = "id_contact") val idContact: Int?,                      //id du contact
         @ColumnInfo(name = "contact_details") val contactDetails: String,           //
         @ColumnInfo(name = "tag") val tag: String,                                  //
         @ColumnInfo(name = "field_position") val fieldPosition: Int                 //
