+++ app/src/main/java/com/example/knocker/model/Contacts.kt
@@ -4,7 +4,7 @@ import androidx.room.*
 
 @Entity(tableName = "contacts_table")
 data class Contacts(
-        @ColumnInfo(name = "id") @PrimaryKey(autoGenerate = true) val id: Long?,    //id du contact
+        @ColumnInfo(name = "id") @PrimaryKey(autoGenerate = true) val id: Int?,    //id du contact
         @ColumnInfo(name = "first_name") val firstName: String,                     //prénom du contact
         @ColumnInfo(name = "last_name") val lastName: String,                       //nom du contact
         @ColumnInfo(name = "profile_picture") val profilePicture: Int,              //image de profile du contact
